scripts/requirements.txt
# Scraping & Parsing
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Gemini AI
google-generativeai>=0.3.0

# Utils
python-dateutil>=2.8.0

